+++ client/src/components/app.tsx
@@ -7,8 +7,9 @@ import { TourProvider } from '@reactour/tour';
 import { newUserSteps } from '../tours/new-user-steps';
 import { useSdUser } from '../contexts/user-context';
 import { MapContextProvider } from '../utils/map-context';
-import { OurGames } from './games/our-games';
 import { TheStiflingDarkPage } from './games/the-stifling-dark/the-stifling-dark-page';
+import { YourMaps } from './map-editor/your-maps';
+import { Games } from './routes/games';
 
 export const App = () => {
 	const { closeSetupGuide } = useSdUser();
@@ -29,8 +30,9 @@ export const App = () => {
 				<Routes>
 					<Route path='/' element={<Home />} />
 					<Route path='/map/:mapId' element={<MapEditor />} />
+					<Route path='/maps' element={<YourMaps />} />
 					<Route path='/maps/new' element={<MapForm />} />
-					<Route path='/games' element={<OurGames />} />
+					<Route path='/games' element={<Games />} />
 					<Route
 						path='/games/the-stifling-dark'
 						element={<TheStiflingDarkPage />}
