+++ client/src/components/routes/games.tsx
@@ -1,0 +1,10 @@
+import { OurGames } from '../games/our-games';
+
+export const Games = () => {
+	return (
+		<div className='content-container'>
+			<h1>Our Games</h1>
+			<OurGames />
+		</div>
+	);
+};
